+++ src/features/express/index.js
@@ -4,21 +4,21 @@ import express from 'express';
 import { call, pipe } from 'ramda';
 import { WEB } from '../../constants';
 import { setHandlerResult, setMetaEnv } from '../../lens/feature';
-import { EXPRESS } from './constants';
-import { getAllRoutes, getExpressConfig } from './lens';
 import useMiddlewares from './util/useMiddlewares';
 import useRoutes from './util/useRoutes';
 import startServer from './util/startServer';
+import { EXPRESS } from './constants';
+import { getAllRoutes, getExpressConfig } from './lens';
 
-const handler = async app => {
-  const config = getExpressConfig(app);
-  const routes = getAllRoutes(app);
-  const api = pipe(call, useMiddlewares, useRoutes(config, routes))(express);
-  const server = http.createServer(api);
+const handler = async root => {
+  const config = getExpressConfig(root);
+  const routes = getAllRoutes(root);
+  const app = pipe(call, useMiddlewares, useRoutes(config, routes))(express);
+  const server = http.createServer(app);
 
   await startServer(config, server);
 
-  return setHandlerResult(api);
+  return setHandlerResult(app);
 };
 
 const Express = pipe(setId(EXPRESS), setMetaEnv(WEB), setHandler(handler));
